pyproject.toml
[tool.poetry]
name = "qywxbot"
version = "0.1.0"
description = ""
authors = ["samzong.lu <samzong.lu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^2.2.2"
pycrypto = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

